Solution to Check Month Days

+++ Check Month Days@@6788dd63490637cfc44e1baa/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main() {
-    float a;
-    scanf("%f",&a);
+    int a;
+    scanf("%d",&a);
     if (a==2){
         printf("28");
     } else if (a==1 || a==3 || a==5 || a==7 || a==8 || a==10|| a==12){
